more work on projectiles, projectile rotation
now with bonus crosshair rotation effect

+++ dueric_summativesdl/PlayerProj.cpp
@@ -3,7 +3,7 @@
 #include <SDL_image.h>
 #include <iostream>
 
-PlayerProj::PlayerProj(int x1, int y1, int x2, int y2, int speed, SDL_Renderer *renderer) : Projectile(x1, y1, x2, y2, speed)
+PlayerProj::PlayerProj(int x1, int y1, int x2, int y2, double speed, SDL_Renderer *renderer) : Projectile(x1, y1, x2, y2, speed)
 {
 	projHeight = 16;
 	projWidth = 6;
